rm soft_argmax_head
