+++ mmpose/models/heads/__init__.py
@@ -3,10 +3,10 @@ from .base_head import BaseHead
 from .heatmap_heads import (CPMHead, HeatmapHead, MSPNHead, SimCCHead,
                             ViPNASHead)
 from .regression_heads import (DSNTHead, IntegralRegressionHead,
-                               RegressionHead, RLEHead, SoftArgmaxHead)
+                               RegressionHead, RLEHead)
 
 __all__ = [
     'BaseHead', 'HeatmapHead', 'CPMHead', 'MSPNHead', 'ViPNASHead',
     'RegressionHead', 'IntegralRegressionHead', 'SimCCHead', 'RLEHead',
-    'DSNTHead', 'SoftArgmaxHead'
+    'DSNTHead'
 ]
